Add Uint64/Int64/Float64/String to ByteConsumer
These methods pull data out of the pool of bytes and turn it into usable
data types. The types are all the largest native type in Go, but you
pass in the number of bits to build the type.

So to get a uint16 you call Uint64(16). This is slightly awkward, but
unifies with the reflect api which works similarly.