Stop filling when we run out of bytes
This allows us to build arbitrary sized recursive datastructures using
Fill().

This seems to suggest that we can build tree like recursive
datastructures. However, because of the depth-first nature of Fill a
binary tree would always fill to one side only and always produce a
leftward (or rightward) skewed unbalanced tree. So recursive
datastructures, are in practical terms, limited to linked list style
datastrucutures. I think this is likely enough.