Add binary tree test
This test clarifies that Fill() will produce a fully unbalanced binary
tree skewed to the left.

This is not a bug, and is currently considered an expected limitation of
Fill(). Providing alternative filling strategies, other that depth
first, would add extra complexity to the package and likely provide
limited use.

If we consider the case of trying to build a binary tree datastructure
for fuzzing purposes. Building such a datastructure with a breadth-first
Fill() implementation would always produce nice balanced trees, but in
fuzzing don't necessarily want nice balanced trees, nor do we want
wildly left skewed trees. We want all the varieties of trees that we can
reasonably produce.

So directly building a tree-like datastructure using Fill() will likely
never produce a useful fuzzing result. At least not without an
extraordinary increase in the implementation complexity of Fill().

If you want to build a tree-like datastructure for fuzzing (a long time
ago I wrote a poor man's fuzzer to test a red-black tree implementation)
we will instead produce a list of steps for building the tree.

The next piece of work will be to create a system to make this easy.