Add fuzz-string-method tag
Also removing the fuzz-string-values tag we just added.

This tag calls a method on the struct to get the []string values.

This is _more_ complex than the previous approach which just had a list
of comma separated values embeded directly in the tag itself. Now you
list the name of a method on the struct which when called returns a
[]string to use.

Simplicity was the chief virtue of the original approach, but there were
some problems which convinced me to abandon it and use the method based
approach instead.

1: Number of string options. If you need a large number of options then
   putting them all in the tag will become cumbersome. This _is_ a
problem, but not the biggest problem IMO.
2: Matching the string values to switch statements in fuzzing code. The
   use-case that I am currently considering for this tag is where we
want to generate a list of many options which are used to drive API
calls in the system under test. In this case you would have to
copy/paste the values from the tag into a switch statement to drive
these different API calls. This will be very error prone. I do not like
this at all. While not perfect, we can now at least maintain a list of
strings to be used in code and not copy/paste between tags and const
declarations.

Time will tell if this is a wise choice.