Explicit depth-first search fill
Here we break down each fill into a step which fills in _one_ value and
returns a slice of the next values which need to be filled (in the form
of a function which does that work).

Simple fill tasks like setting a string or an integer return no new fill
functions. But complex fill tasks like creating a map and then filling
it with values will create the map, and return functions for filling in
the values.

Here we iterate over those values in a depth first manner, which
precisely matches how we did it using just recursive function calls
previously.

Next step will be to switch to breadth first. Which will break a lot of
tests :(