Switch to breadth first filling
This change is best characterised by the binary tree test. This test has
changed from demonstrating that the fill algorithm will create an
unbalanced left skewed tree, to showing a nicely balanced binary tree.

There were some subtle problems with this change. When filling in a map
whose values are a struct value (i.e. not a pointer to a struct type) we
would first create the struct, add it to the map and _then_ fill in the
structs fields. This was a natural result of performing the fill breadth
first. But the result is that the struct is always unfilled when it is
added to the map, so the filled values are never observed.

To solve this problem we now fill all types eagerly (i.e. by directly
calling fill rather than creating a fill callback function) until we
reach a pointer. Pointer types are created and then filling the
allocated pointer type is deferred by creating a fill callback. This
satisfies our need for a breadth-first fill algorithm, with the detail
that structs, slices and maps are all filled eagerly as a 'fill unit'.