Look up pointer and value receiver methods
This makes the method lookup more reliable.

In the future many of the debug print statements will turn into errors
which should stop the test from running. As it is, if you make a mistake
in your tags the test will run and not do what you expected. This feels
like an extremely bad user-experience, the case where your tests are
passing but not actually testing what you think is a nightmare.