Adding stack fuzz example
Here we implement a stack implementation which passes a small set of
reasonable looking tests. The stack implementation has a problem where
the integer which tracks the top position of the stack is only 16 bits
wide and will overflow under fairly normal usage conditions. However the
conditions are difficult to hit in conventional unit tests.

We add a fuzzing test which will reliably uncover the overflow failure
condition. A user should be able to run

go test -fuzz=Fuzz

and see a failing case very quickly.

Adding comments to fuzz_test.go for stack