Adding sortedtree example
In this example we build a simple tree datastructure for sorting
strings. It has a fatal flaw that when comparing strings it only
compares the first 64 characters. Our unit tests pass all their tests,
because they have relatively short test string values.

We create a fuzz test which builds a tree with random input strings and
_should_ find the bug.

This test demonstrates a different fuzzing strategy where we use a
reference data structure (in this case a sorted slice of strings) to
compare against the result produced by the data structure under test.