Partial implementation of visitor pattern
We have the visitValue function which looks at the kind of a value and
chooses the correct callback to invoke. This is fine and good.

However, when we look at the struct callback (as an example) we can see
recursive calls back into visitValue. This means that the way in which
the structures are being traversed and processed is partly determined by
the visitValue function, and partly by the visitor implementation. This
is very bad and not good.

The next step will be to move the processing of complex recursive types,
channel, slice, array, map and struct into the visitValue function. This
means that each visitor implementation will be considerably simpler and
have no real way to alter the order in which types are processed.