Create proper valuePath representation
This type contains both the user-friendly names for types along the path
as well as the reflect.Values for them. This means that we can build
nice human strings but also ask questions like has this type already
appeared on this path, to detect recursion.

We also fixed up the array description along the way.