Allow descriptions of recursive types
Previously this wouldn't work because the vistor would keep visiting
recursive structs, since there is no data source to exhaust there is no
end to this. We now allow the visitor callback to choose to stop
processing of a struct.

We also now describe a struct if it is found to be recursive. A
non-recursive struct is described via its fields, but since we stop
processing a recursive struct we must describe it before we abandon it.