Allow Describe() to work on root slice types
Previously the describe visitor couldn't stop the root slice from
growing. And the root slice would continue to grow so long as there was
data to build with, and the describe visitor doesn't consume data.

Now we can short circuit the root slice growth after adding the first
element, allowing the describe visitor to work properly.