Add 'fuzz-interface-method' tag
This tag identifies a method which is used to generate a list of
concrete types (wrapped in the any interface) which can be used to
instantiate the field it is attached to.

This should allow us to actually instantiate interface fields. But I
suspect there are some details to be worked out here.