Apply slice of types in visitInterface() method
Previously visitInterface() was a no-op for the fillVisitor. It just
ignored interfaces. Now we look to see if there is a slice of interface
values provided via the 'fuzz-interface-method' tag and choose on of
those if available.

The visitInterface() method now returns true if a value was set in the
interface reflect.Value, in this case this new value needs to be visited
to fill in its fields or simply set it value. Returning false means no
value was set in the interface and it should not be visited.

The slice used here is a slice of 'any', but is used as a list of
concrete types that are used to build new values of that type. This
means that the actual values in the slice are never used to set an
interface directly. So they are safe to resuse.

We allow the setting of interfaces which are more complex (constrained)
than just 'any'. However, we do restrict all types to pointer-types.
This choice was made because a value type in an interface cannot be set
after extracting it from behind the interface via reflection. It's
likely there is a solution to this, but it would add greater complexity
to the codebase. A restriction to _only_ pointer types behind an
interface is likely to be a non-problem. We can relax this restriction
in the future if it feels like a problem.