Add FillSlice[T] convenience function
This function allows the caller to pass in a slice of allowable types
([]T) and the source of random byte. We will then generate and fill a
slice by selecting from those allowable types. This is exactly the same
as using the `fuzz-interface-method` annotation, but in a more readable
and straight-forward format of just a function taking arguments.