Allow unbounded slices
Slices which do not have a min/max tag attached are unbounded in size.
We build them recursively, by increasing the length of the slice each
time we visit, and appending a recursive visitFunc to revisit the slice
again. This is added _after_ the processing has completed for the new
element added to the slice. Eventually we run out of data and will stop
appending new elements to the slice.

Some carefuly handling is required to avoid infinite recursion when
describing a slice.

We will make the same change for maps in the future.