Allow interface value methods for specific types
Previously the way we processed method tags for interface values
unintentionally forced the interface methods to return []any or have
their values ignored. This is because we checked that the type of the
slice returned matched the type that the method tag was storing. In the
case of interface value methods this type is []any.

Now we manually copy the returned slice, which can be of a specific
interface type. In the process this _does_ have the effect of converting
the restricted interface type to any, so we lose type information in the
process. However, we check that the types match the type required by the
field when assigning values from the interface slice.