Require exact types from fuzz-method tag methods
Previously we required that fuzz-method methods were of the widest type
for their type category i.e.

int8/16/32/64 fields required a method returning []int64
uint8/16/32/64 fields required []uint64
float8/16/32/64 fields required []float64
interface fields required []any
string fields required []string

Now we require a slice of the exact type of the field that will be
assigned to. This is somewhat tricky because we still store the tag
values in a slice of the widest type. So now we carefully check that
each slice element, returned by the fuzz-method methd, can be assigned
as-is and then convert it to the wider type. When the value is later
assigned to its field the type is narrowed if necessary. Without loss of
information.

This is a bit tricky, or at least feels like a breeding ground for
subtle bugs. So we have created a separate test which exhaustively tests
all type conversions (although it does not test a wide range of values).