Panic when fuzz-method methods are badly set up
Previously if we found problems getting the slice of values from a
fuzz-method method we just gave up and silently set the tag to not-set.
This is likely going to cause problems for people with incorrectly set
up fuzz-method tags, who think they are correct and are just running
tests will missing data values.

Now we panic on any misconfiguration with a helpful message. We have
written up a full set of tests to capture all of these failure modes.